+++ pico/universal/common/src/reference.c
@@ -1,4 +1,0 @@
-#include <stdint.h>
-
-void dispatch(uint32_t v) {
-}