+++ pico/universal/picow/wifi+usb/src/universal.c
@@ -2,32 +2,35 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "hardware/gpio.h"
-#include "hardware/rtc.h"
-#include "hardware/watchdog.h"
+#include <hardware/gpio.h>
+#include <hardware/rtc.h>
+#include <hardware/watchdog.h>
 
-#include "pico/binary_info.h"
-#include "pico/multicore.h"
-#include "pico/stdlib.h"
-#include "pico/util/datetime.h"
+#include <pico/binary_info.h>
+#include <pico/multicore.h>
+#include <pico/stdlib.h>
+#include <pico/util/datetime.h>
 
-#include "TPIC6B595.h"
-#include "acl.h"
-#include "buzzer.h"
-#include "common.h"
-#include "led.h"
-#include "logd.h"
-#include "read.h"
-#include "relays.h"
-#include "sdcard.h"
-#include "sys.h"
-#include "uart.h"
-#include "wiegand.h"
-#include "write.h"
+#include <TPIC6B595.h>
+#include <acl.h>
+#include <buzzer.h>
+#include <common.h>
+#include <led.h>
+#include <logd.h>
+#include <picow.h>
+#include <read.h>
+#include <relays.h>
+#include <sdcard.h>
+#include <sys.h>
+#include <tcpd.h>
+#include <uart.h>
+#include <usb.h>
+#include <wiegand.h>
+#include <write.h>
 
-#include "../include/reference.h"
+#include "../include/universal.h"
 
-#define VERSION "v0.8.7"
+#define VERSION "v0.8.5"
 
 // GPIO
 const uint32_t MSG = 0xf0000000;
@@ -35,13 +38,12 @@ const uint32_t MSG_WATCHDOG = 0x00000000;
 const uint32_t MSG_SYSCHECK = 0x10000000;
 const uint32_t MSG_RX = 0x20000000;
 const uint32_t MSG_TX = 0x30000000;
-const uint32_t MSG_CARD = 0x40000000;
-const uint32_t MSG_CODE = 0x50000000;
 const uint32_t MSG_KEYPAD_DIGIT = 0x60000000;
 const uint32_t MSG_LED = 0x70000000;
 const uint32_t MSG_RELAY = 0x80000000;
 const uint32_t MSG_DOOR = 0x90000000;
 const uint32_t MSG_PUSHBUTTON = 0xa0000000;
+const uint32_t MSG_TCPD_POLL = 0xc0000000;
 const uint32_t MSG_LOG = 0xd0000000;
 const uint32_t MSG_SYSINIT = 0xe0000000;
 const uint32_t MSG_DEBUG = 0xf0000000;
@@ -49,6 +51,7 @@ const uint32_t MSG_DEBUG = 0xf0000000;
 // FUNCTION PROTOTYPES
 
 void setup_gpio(void);
+void setup_cyw43(void);
 void sysinit();
 
 // GLOBALS
@@ -62,7 +65,7 @@ card last_card = {
 };
 
 int main() {
-    bi_decl(bi_program_description("Pico-Wiegand interface"));
+    bi_decl(bi_program_description("Pico-Wiegand reference interface (USB+WIFI)"));
     bi_decl(bi_program_version_string(VERSION));
 
     stdio_init_all();
@@ -73,10 +76,13 @@ int main() {
     rtc_init();
     sleep_us(64);
 
-    // ... initialise FIFO, UART and timers
+    // ... initialise FIFO, USB and timers
     queue_init(&queue, sizeof(uint32_t), 64);
-    setup_uart();
+    setup_usb();
     alarm_pool_init_default();
+
+    // ... initialise CYW43
+    setup_cyw43();
 
     // ... initialise reader/emulator
     add_alarm_in_ms(250, startup, NULL, true);
@@ -152,7 +158,12 @@ int main() {
         if ((v & MSG) == MSG_LOG) {
             char *b = (char *)(SRAM_BASE | (v & 0x0fffffff));
             printf("%s", b);
+            tcpd_log(b);
             free(b);
+        }
+
+        if ((v & MSG) == MSG_TCPD_POLL) {
+            tcpd_poll();
         }
 
         if ((v & MSG) == MSG_DEBUG) {
@@ -203,7 +214,7 @@ void sysinit() {
     static repeating_timer_t syscheck_rt;
 
     if (!initialised) {
-        puts("                     *** WIEGAND REFERENCE IMPLEMENTATION");
+        puts("                     *** WIEGAND REFERENCE IMPLEMENTATION (USB)");
 
         if (!gpio_get(JUMPER_READ) && gpio_get(JUMPER_WRITE)) {
             mode = READER;
@@ -213,12 +224,14 @@ void sysinit() {
             mode = UNKNOWN;
         }
 
-        sdcard_initialise(mode, true);
+        logd_initialise(mode);
+        sdcard_initialise(mode, false); // GPIO interrupt conflict with SD card detect
         read_initialise(mode);
         write_initialise(mode);
         led_initialise(mode);
         buzzer_initialise(mode);
         TPIC_initialise(mode);
+        tcpd_initialise(mode);
         acl_initialise();
 
         if (!relay_initialise(mode)) {